Modifies presentation logic of RemoveRoleUiHandler.

+++ app/src/main/java/com/ecc/hibernate_xml/ui_handler/person/modify/role/RemoveRoleUiHandler.java
@@ -1,8 +1,8 @@
 package com.ecc.hibernate_xml.ui_handler.person.modify.role;
 
 import java.util.stream.Collectors;
+import java.util.List;
 
-import com.ecc.hibernate_xml.ui_handler.CompositeUiHandler;
 import com.ecc.hibernate_xml.ui_handler.UiHandler;
 import com.ecc.hibernate_xml.util.InputHandler;
 import com.ecc.hibernate_xml.service.PersonService;
@@ -11,7 +11,6 @@ import com.ecc.hibernate_xml.model.Person;
 import com.ecc.hibernate_xml.model.Role;
 
 public class RemoveRoleUiHandler extends UiHandler {
-
 	private static final String PROMPT = "Please choose role to remove from the following: ";
 
 	private PersonService personService;
@@ -27,17 +26,26 @@ public class RemoveRoleUiHandler extends UiHandler {
 
 	@Override 
 	public void onHandle() throws Exception {
+		List<Role> roles = roleService.listRoles(person);
 
-		String listOfRoles = roleService.listRoles(person).stream()
-			.map(role -> String.format("[ID=%d] %s", role.getId(), role.getName()))
-			.collect(Collectors.joining("\n"));
+		System.out.println("-------------------");
+		if (roles.isEmpty()) {
+			System.out.println("There are no assigned roles to remove.");
+		}
+		else {		
+			String listOfRoles = roles.stream()
+				.map(role -> role.toString())
+				.collect(Collectors.joining("\n"));
 
-		Integer roleId = InputHandler.getNextLine(
-			String.format("%s\n%s\nRole ID: ", PROMPT, listOfRoles), Integer::valueOf);
+			Integer roleId = InputHandler.getNextLine(
+				String.format("%s\n%s\nRole ID: ", PROMPT, listOfRoles), Integer::valueOf);
 
-		personService.removeRoleFromPerson(roleId, person);
-		System.out.println(String.format(
-			"Successfully removed role ID \"%d\" to Person \"%s\"!", roleId, person.getName()));
+			personService.removeRoleFromPerson(roleId, person);
+			System.out.println(String.format(
+				"Successfully removed role ID \"%d\" to Person ID [%d] \"%s\"!", roleId, 
+				person.getId(), person.getName()));
+		}
+		System.out.println("-------------------");
 	}
 
 	@Override
